move session creation into session registry
